agent/pyproject.toml
[project]
name = "agent"
version = "0.1.0"
description = ""
authors = [
    {name = "Eric Du",email = "ericdu117@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai (>=2.16.0,<3.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
